move lifecycle manager to correct package.
